requirements.txt
torch==2.2.0
torchvision==0.17.0
transformers==4.39.0
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.3
scikit-learn==1.4.0
nltk==3.8.1
tqdm==4.66.2
pillow==10.2.0
medspacy>=0.2.0.0
rouge-score==0.1.2
bert-score==0.3.13
tensorboard==2.15.1
opencv-python==4.9.0.80
wandb==0.16.3 